SendMultipleEmailsConcurrently function moved to mail.go
