+++ occupi-backend/pkg/mail/mail.go
@@ -1,6 +1,8 @@
 package mail
 
 import (
+	"sync"
+
 	"github.com/COS301-SE-2024/occupi/occupi-backend/configs"
 	"gopkg.in/gomail.v2"
 )
@@ -25,4 +27,28 @@ func SendMail(to string, subject string, body string) error {
 	}
 
 	return nil
+}
+
+func SendMultipleEmailsConcurrently(emails []string, subject, body string) []string {
+	// Use a WaitGroup to wait for all goroutines to complete
+	var wg sync.WaitGroup
+	var emailErrors []string
+	var mu sync.Mutex
+
+	for _, email := range emails {
+		wg.Add(1)
+		go func(email string) {
+			defer wg.Done()
+			if err := SendMail(email, subject, body); err != nil {
+				mu.Lock()
+				emailErrors = append(emailErrors, email)
+				mu.Unlock()
+			}
+		}(email)
+	}
+
+	// Wait for all email sending goroutines to complete
+	wg.Wait()
+
+	return emailErrors
 }
